requirements.txt
discord
ruamel.yaml
requests
numpy
grilops